refactor(Controller): 인덱스와 게임의 RestController 분리
